refactor: remove list verb * after talking with jsbroks, the command should work similar to   kubectl, I've moved the list verb under go, will refactor to dry up   the code as most things are the same beetween the calls
