+++ cmd/ctrlc/root/api/api.go
@@ -5,7 +5,6 @@ import (
 	"os"
 
 	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/get"
-	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/list"
 	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/upsert"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -46,7 +45,6 @@ func NewAPICmd() *cobra.Command {
 
 	cmd.AddCommand(get.NewGetCmd())
 	cmd.AddCommand(upsert.NewUpsertCmd())
-	cmd.AddCommand(list.NewListCmd())
 
 	return cmd
 }
