+++ cmd/ctrlc/root/api/get/deployments/deployments.go
@@ -1,4 +1,4 @@
-package environments
+package deployments
 
 import (
 	"fmt"
@@ -9,14 +9,14 @@ import (
 	"github.com/spf13/viper"
 )
 
-func NewListEnvironmentsCmd() *cobra.Command {
+func NewDeploymentsCmd() *cobra.Command {
 	var limit int
 	var offset int
 
 	cmd := &cobra.Command{
-		Use:   "environments",
-		Short: "List environments",
-		Long:  `Commands for getting environments.`,
+		Use:   "deployments",
+		Short: "List deployments",
+		Long:  `Commands for getting deployments.`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			apiURL := viper.GetString("url")
 			apiKey := viper.GetString("api-key")
@@ -29,17 +29,17 @@ func NewListEnvironmentsCmd() *cobra.Command {
 
 			workspaceID := client.GetWorkspaceID(cmd.Context(), workspace)
 
-			params := &api.ListEnvironmentsParams{}
+			params := &api.ListDeploymentsParams{}
 			if limit > 0 {
 				params.Limit = &limit
 			}
 			if offset > 0 {
 				params.Offset = &offset
 			}
-			resp, err := client.ListEnvironments(cmd.Context(), workspaceID.String(), params)
+			resp, err := client.ListDeployments(cmd.Context(), workspaceID.String(), params)
 			if err != nil {
 
-				return fmt.Errorf("failed to get environments: %w", err)
+				return fmt.Errorf("failed to get deployments: %w", err)
 			}
 
 			return cliutil.HandleResponseOutput(cmd, resp)
