+++ cmd/ctrlc/root/api/get/get.go
@@ -1,7 +1,10 @@
 package get
 
 import (
+	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/get/deployments"
+	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/get/environments"
 	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/get/resources"
+	"github.com/ctrlplanedev/cli/cmd/ctrlc/root/api/get/systems"
 	"github.com/spf13/cobra"
 )
 
@@ -16,6 +19,9 @@ func NewGetCmd() *cobra.Command {
 	}
 
 	cmd.AddCommand(resources.NewResourcesCmd())
+	cmd.AddCommand(deployments.NewDeploymentsCmd())
+	cmd.AddCommand(systems.NewSystemsCmd())
+	cmd.AddCommand(environments.NewEnvironmentsCmd())
 
 	return cmd
 }
